新增： 快速构建树的工具类 权限、角色相关接口 微信登录demo 国际化返回

+++ sso-mybatis/src/main/java/com/zhanglx/sso/mybatis/handler/DefaultDBFieldHandler.java
@@ -32,7 +32,7 @@ public class DefaultDBFieldHandler implements MetaObjectHandler {
             this.strictInsertFill(metaObject, "createTime", LocalDateTime.class, current);
             this.strictInsertFill(metaObject, "updateBy", Long.class, userId);
             this.strictInsertFill(metaObject, "updateTime", LocalDateTime.class, current);
-            this.strictInsertFill(metaObject, "delFlag", Integer.class, 0);
+            this.strictInsertFill(metaObject, "delFlag", Long.class, 0L);
         }
     }
